+++ services/vue/src/components/private/editor/aggregation/Form.vue
@@ -53,38 +53,68 @@
             </v-col>
           </v-row>
 
-          <v-row v-if="aggregation.type">
-            <v-col>
-              <v-combobox
-                v-model="aggregation.field"
-                :label="$t('$ezreeport.editor.aggregation.field')"
-                :items="fieldOptions"
-                :rules="[(v) => !!v || $t('$ezreeport.required')]"
-                :readonly="readonly"
-                :disabled="disabled"
-                :return-object="false"
-                prepend-icon="mdi-form-textbox"
-                variant="underlined"
-                required
-              />
-            </v-col>
-          </v-row>
+          <v-slide-y-transition>
+            <v-row v-if="aggregation.type">
+              <v-col>
+                <v-combobox
+                  v-model="aggregation.field"
+                  :label="$t('$ezreeport.editor.aggregation.field')"
+                  :items="fieldOptions"
+                  :rules="[(v) => !!v || $t('$ezreeport.required')]"
+                  :readonly="readonly"
+                  :disabled="disabled"
+                  :return-object="false"
+                  prepend-icon="mdi-form-textbox"
+                  variant="underlined"
+                  required
+                />
+              </v-col>
+            </v-row>
+          </v-slide-y-transition>
 
-          <v-row v-if="isMetric === false && aggregation.type !== 'date_histogram'">
-            <v-col>
-              <v-text-field
-                :model-value="`${aggregation.size ?? 10}`"
-                :label="$t('$ezreeport.editor.aggregation.size')"
-                :readonly="readonly"
-                :disabled="disabled"
-                type="number"
-                prepend-icon="mdi-image-size-select-small"
-                variant="underlined"
-                hide-details
-                @update:model-value="aggregation.size = Number.parseInt($event)"
-              />
-            </v-col>
-          </v-row>
+          <v-slide-y-transition>
+            <v-row v-if="isMetric === false && aggregation.type !== 'date_histogram'">
+              <v-col cols="12">
+                <v-text-field
+                  :model-value="`${aggregation.size ?? 10}`"
+                  :label="$t('$ezreeport.editor.aggregation.size')"
+                  :readonly="readonly"
+                  :disabled="disabled"
+                  type="number"
+                  prepend-icon="mdi-image-size-select-small"
+                  variant="underlined"
+                  hide-details
+                  @update:model-value="aggregation.size = Number.parseInt($event)"
+                />
+              </v-col>
+
+              <v-col cols="6">
+                <v-switch
+                  v-model="showMissing"
+                  :label="$t('$ezreeport.editor.aggregation.missing:show')"
+                  :disabled="readonly"
+                  prepend-icon="mdi-progress-question"
+                  color="primary"
+                  hide-details
+                />
+              </v-col>
+
+              <v-col cols="6">
+                <v-slide-x-transition>
+                  <v-text-field
+                    v-if="showMissing"
+                    v-model="aggregation.missing"
+                    :label="$t('$ezreeport.editor.aggregation.missing:label')"
+                    :readonly="readonly"
+                    :disabled="disabled"
+                    prepend-icon="mdi-tooltip-question-outline"
+                    variant="underlined"
+                    hide-details
+                  />
+                </v-slide-x-transition>
+              </v-col>
+            </v-row>
+          </v-slide-y-transition>
         </template>
 
         <slot name="text" />
@@ -98,7 +128,7 @@ import {
   type InnerAggregation,
   isRawAggregation,
   isBaseAggregation,
-  aggregationFieldType,
+  aggregationFieldTypes,
 } from '~/lib/aggregations';
 import { type FigureAggregation, type AggregationType, aggregationTypes } from '~sdk/helpers/aggregations';
 
@@ -157,14 +187,29 @@ const isMetric = computed(() => {
   }
   return aggDef.type === 'metric';
 });
+/** If we should show the missing values */
+const showMissing = computed({
+  get: () => {
+    if (isRawAggregation(aggregation.value)) {
+      return false;
+    }
+    return !!aggregation.value.missing;
+  },
+  set: (value) => {
+    if (isRawAggregation(aggregation.value)) {
+      return;
+    }
+    aggregation.value.missing = value ? 'Missing' : undefined;
+  },
+});
 /** Type of fields needed for the current aggregation */
-const fieldType = computed(() => (
+const fieldTypes = computed(() => (
   !isRawAggregation(aggregation.value) && aggregation.value.type
-    ? aggregationFieldType.get(aggregation.value.type)
+    ? aggregationFieldTypes.get(aggregation.value.type)
     : undefined
 ));
 /** Options for the field, based on current mapping */
-const fieldOptions = computed(() => getOptionsFromMapping(fieldType.value, { dateField: true }));
+const fieldOptions = computed(() => getOptionsFromMapping(fieldTypes.value, { dateField: true }));
 /** Options for the aggregation type */
 const typeOptions = computed(() => {
   let types = [...aggregationTypes];