+++ services/vue/src/components/private/task/EditionForm.vue
@@ -80,6 +80,19 @@
               />
 
               <EditorFilterList v-model="filters" />
+
+              <v-btn
+                v-if="showAdvanced"
+                v-tooltip:top="$t('$ezreeport.task.superUserMode:tooltip')"
+                :text="$t('$ezreeport.task.superUserMode')"
+                prepend-icon="mdi-tools"
+                append-icon="mdi-tools"
+                color="warning"
+                variant="flat"
+                block
+                class="mt-4"
+                @click="emit('open:advanced', task)"
+              />
             </template>
           </v-expansion-panel>
         </v-expansion-panels>
@@ -92,7 +105,7 @@
       <slot name="actions" />
 
       <v-btn
-        :text="$t('$ezreeport.confirm')"
+        :text="$t('$ezreeport.save')"
         :disabled="!isValid"
         append-icon="mdi-pencil"
         color="primary"
@@ -119,12 +132,16 @@ const props = defineProps<{
   modelValue: Task,
   /** Should show namespace */
   showNamespace?: boolean,
+  /** Should show advanced button */
+  showAdvanced?: boolean
 }>();
 
 // Components events
 const emit = defineEmits<{
   /** Updated task */
   (e: 'update:modelValue', value: Task): void
+  /** Asked to open task in advanced form */
+  (e: 'open:advanced', value: InputTask): void
 }>();
 
 // Utils composables
@@ -135,8 +152,6 @@ const { refreshMapping } = useTemplateEditor();
 const isValid = ref(false);
 /** Are namespaces loading */
 const loadingNamespaces = ref(false);
-/** Current namespace */
-const namespace = ref<Omit<Namespace, 'fetchLogin' | 'fetchOptions'> | undefined>();
 /** Task to create */
 const task = ref<InputTask>({
   name: props.modelValue.name,
@@ -152,6 +167,7 @@ const task = ref<InputTask>({
   namespace: props.modelValue.namespace,
 });
 
+/** Filters of task */
 const filters = computed({
   get: () => new Map((task.value.template.filters ?? []).map((f) => [f.name, f])),
   set: (v) => {
@@ -162,6 +178,29 @@ const filters = computed({
     }
     task.value.template.filters = undefined;
   },
+});
+/** Is form namespaced */
+const isNamespaced = computed(() => !props.showNamespace);
+/** Current namespace's id */
+const namespaceId = computed(() => task.value.namespaceId);
+/** Curent namespace */
+const namespace = computedAsync(async () => {
+  let value: Omit<Namespace, 'fetchLogin' | 'fetchOptions'> | undefined;
+
+  if (isNamespaced.value) {
+    return value;
+  }
+
+  loadingNamespaces.value = true;
+  try {
+    const currentNamespaces = await getCurrentNamespaces();
+    value = currentNamespaces.find((n) => n.id === namespaceId.value);
+  } catch (e) {
+    handleEzrError(t('$ezreeport.task.errors.fetchNamespaces'), e);
+  }
+  loadingNamespaces.value = false;
+
+  return value;
 });
 
 function onTargetUpdated(targets: string | string[] | undefined) {
@@ -185,17 +224,6 @@ function onTargetUpdated(targets: string | string[] | undefined) {
   );
 }
 
-async function refreshNamespace() {
-  loadingNamespaces.value = true;
-  try {
-    const currentNamespaces = await getCurrentNamespaces();
-    namespace.value = currentNamespaces.find((n) => n.id === task.value.namespaceId);
-  } catch (e) {
-    handleEzrError(t('$ezreeport.errors.refreshNamespaces'), e);
-  }
-  loadingNamespaces.value = false;
-}
-
 async function save() {
   try {
     const result = await upsertTask({ ...task.value, id: props.modelValue.id });
@@ -204,9 +232,5 @@ async function save() {
   } catch (e) {
     handleEzrError(t('$ezreeport.task.errors.update'), e);
   }
-}
-
-if (props.showNamespace) {
-  refreshNamespace();
 }
 </script>