+++ services/vue/src/components/public/TaskPresetTable.vue
@@ -9,12 +9,16 @@
   >
     <template #top>
       <v-toolbar
-        :title="`${titlePrefix || ''}${$t('$ezreeport.task-preset.title:list', total)}`"
+        :title="title"
         color="transparent"
         density="comfortable"
       >
         <template v-if="$slots.prepend" #prepend>
           <slot name="prepend" />
+        </template>
+
+        <template v-if="$slots.title" #title>
+          <slot name="title" :title="title" />
         </template>
 
         <template #append>
@@ -203,8 +207,15 @@ import {
 type VDataTableHeaders = Exclude<VDataTable['$props']['headers'], undefined>;
 
 // Components props
-defineProps<{
+const props = defineProps<{
   titlePrefix?: string;
+  itemsPerPageOptions?: number[] | { title: string; value: number }[];
+  itemsPerPage?: number;
+}>();
+
+// Components events
+const emit = defineEmits<{
+  (e: 'update:itemsPerPage', value: number): void
 }>();
 
 // Utils composable
@@ -215,6 +226,8 @@ const selectedTaskPresets = ref<TaskPreset[]>([]);
 const updatedTaskPreset = ref<TaskPreset | undefined>();
 const isFormOpen = ref(false);
 
+/** Items per page shortcut */
+const itemsPerPage = computed({ get: () => props.itemsPerPage || 10, set: (v) => emit('update:itemsPerPage', v) });
 /** List of templates */
 const {
   total,
@@ -224,8 +237,12 @@ const {
   vDataTableOptions,
 } = useServerSidePagination(
   (params) => getAllTaskPresets(params),
-  { sortBy: 'name', include: ['template.tags', 'template.hidden'] },
+  {
+    sortBy: 'name', include: ['template.tags', 'template.hidden'], itemsPerPage, itemsPerPageOptions: props.itemsPerPageOptions,
+  },
 );
+
+const title = computed(() => `${props.titlePrefix || ''}${t('$ezreeport.task-preset.title:list', total.value)}`);
 
 /** Headers for table */
 const headers = computed((): VDataTableHeaders => [